+++ hutool-core/src/main/java/cn/hutool/core/util/CreditCodeUtil.java
@@ -100,7 +100,7 @@ public class CreditCodeUtil {
 
 		//
 		for (int i = 0; i < 2; i++) {
-			int num = RandomUtil.randomInt(BASE_CODE_ARRAY.length - 1);
+			int num = RandomUtil.randomInt(BASE_CODE_ARRAY.length);
 			buf.append(Character.toUpperCase(BASE_CODE_ARRAY[num]));
 		}
 		for (int i = 2; i < 8; i++) {
@@ -108,7 +108,7 @@ public class CreditCodeUtil {
 			buf.append(BASE_CODE_ARRAY[num]);
 		}
 		for (int i = 8; i < 17; i++) {
-			int num = RandomUtil.randomInt(BASE_CODE_ARRAY.length - 1);
+			int num = RandomUtil.randomInt(BASE_CODE_ARRAY.length);
 			buf.append(BASE_CODE_ARRAY[num]);
 		}