+++ hutool-core/src/main/java/cn/hutool/core/util/URLUtil.java
@@ -807,8 +807,9 @@ public class URLUtil extends URLEncodeUtil {
 		} else {
 			// 如果资源打在jar包中或来自网络，使用网络请求长度
 			// issue#3226, 来自Spring的AbstractFileResolvingResource
+			URLConnection con = null;
 			try {
-				final URLConnection con = url.openConnection();
+				con = url.openConnection();
 				useCachesIfNecessary(con);
 				if (con instanceof HttpURLConnection) {
 					final HttpURLConnection httpCon = (HttpURLConnection) con;
@@ -817,6 +818,10 @@ public class URLUtil extends URLEncodeUtil {
 				return con.getContentLengthLong();
 			} catch (final IOException e) {
 				throw new IORuntimeException(e);
+			} finally {
+				if (con instanceof HttpURLConnection) {
+					((HttpURLConnection) con).disconnect();
+				}
 			}
 		}
 	}