+++ CHANGELOG.md
@@ -24,6 +24,9 @@
 * 【core   】      修复`Validator.isBetween`在高精度Number类型下存在精度丢失问题（pr#4136@Github）
 * 【core   】      修复`FileNameUtil.extName`在特殊后缀判断逻辑过于宽松导致误判问题（pr#4142@Github）
 * 【core   】      修复`TypeUtil.getClass`无法识别`GenericArrayType`问题（pr#4138@Github）
+* 【core   】      修复`CreditCodeUtil.randomCreditCode`部分字母未使用问题（pr#4149@Github）
+* 【core   】      修复`CacheableAnnotationAttribute`可能并发问题（pr#4149@Github）
+* 【core   】      修复`URLUtil.url`未断开连接问题（pr#4149@Github）
 
 -------------------------------------------------------------------------------------------------------------
 # 5.8.41(2025-10-12)